갤러리 화면 수정 및 검색 화면 수정 파일 위치 이동
