+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/local/GalleryLocalDataSourceImpl.kt
@@ -289,8 +289,8 @@ class GalleryLocalDataSourceImpl @Inject constructor(
             .map { (albumName, galleryDataList) ->
                 GalleryAlbumData(
                     thumbnail = galleryDataList.firstOrNull()?.mediaData?.toUri() ?: Uri.EMPTY,
-                    name = albumName,
-                    count = galleryDataList.size,
+                    folderName = albumName,
+                    folderFileCount = galleryDataList.size,
                 )
             }
     }
