+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/model/gallery/GalleryAlbumData.kt
@@ -4,6 +4,6 @@ import android.net.Uri
 
 class GalleryAlbumData(
     val thumbnail: Uri,
-    val name: String,
-    val count: Int
+    val folderName: String,
+    val folderFileCount: Int
 )
