+++ app/src/main/java/com/dnd_9th_3_android/gooding/data/model/search/RecentKeywordData.kt
@@ -2,5 +2,5 @@ package com.dnd_9th_3_android.gooding.data.model.search
 
 data class RecentKeywordData(
     val id: Int,
-    val content: Int,
+    val content: String,
 )
