+++ app/src/main/java/com/dnd_9th_3_android/gooding/presentation/gallery/GalleryAlbumUiData.kt
@@ -4,6 +4,6 @@ import android.net.Uri
 
 data class GalleryAlbumUiData(
     val thumbnail: Uri,
-    val name: String,
-    val count: Int
+    val folderName: String,
+    val folderFileCount: Int
 )
