+++ app/src/main/java/com/dnd_9th_3_android/gooding/presentation/search_feed/SearchFeedListActivity.kt
@@ -237,125 +237,3 @@ class SearchFeedListActivity : AppCompatActivity() {
             Intent(context, SearchFeedListActivity::class.java)
     }
 }
-
-class RecentKeywordListAdapter(
-    private val onClick: (RecentKeywordData) -> Unit
-) : ListAdapter<RecentKeywordData, RecentKeywordListAdapter.RecentFeedViewHolder>(
-    RecentKeywordDiffUtil
-) {
-
-    class RecentFeedViewHolder(
-        private val binding: ItemRecentKeywordBinding,
-        private val onClick: (RecentKeywordData) -> Unit
-    ) : RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(recentKeywordData: RecentKeywordData) {
-            binding.root.setOnClickListener {
-                onClick(recentKeywordData)
-            }
-
-            binding.ivDelete.setOnClickListener {
-
-            }
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecentFeedViewHolder {
-        val binding =
-            ItemRecentKeywordBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return RecentFeedViewHolder(binding, onClick)
-    }
-
-    override fun onBindViewHolder(holder: RecentFeedViewHolder, position: Int) {
-        val recentKeyword = getItem(position)
-        holder.bind(recentKeyword)
-    }
-
-    companion object RecentKeywordDiffUtil : DiffUtil.ItemCallback<RecentKeywordData>() {
-        override fun areItemsTheSame(oldItem: RecentKeywordData, newItem: RecentKeywordData): Boolean {
-            return oldItem.id == newItem.id
-        }
-
-        override fun areContentsTheSame(oldItem: RecentKeywordData, newItem: RecentKeywordData): Boolean {
-            return oldItem == newItem
-        }
-    }
-}
-
-class PopularKeywordListAdapter(
-    private val onClick: (PopularKeywordData) -> Unit
-) : ListAdapter<PopularKeywordData, PopularKeywordListAdapter.PopularKeywordViewHolder>(
-    PopularKeywordDiffUtil
-) {
-
-    class PopularKeywordViewHolder(
-        private val binding: ItemPopularKeywordBinding,
-        private val onClick: (PopularKeywordData) -> Unit
-    ) : RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(popularKeywordData: PopularKeywordData) {
-            binding.root.setOnClickListener {
-                onClick(popularKeywordData)
-            }
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PopularKeywordViewHolder {
-        val binding =
-            ItemPopularKeywordBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return PopularKeywordViewHolder(binding, onClick)
-    }
-
-    override fun onBindViewHolder(holder: PopularKeywordViewHolder, position: Int) {
-        val popularKeyword = getItem(position)
-        holder.bind(popularKeyword)
-    }
-
-    companion object PopularKeywordDiffUtil : DiffUtil.ItemCallback<PopularKeywordData>() {
-        override fun areItemsTheSame(oldItem: PopularKeywordData, newItem: PopularKeywordData): Boolean {
-            return oldItem.id == newItem.id
-        }
-
-        override fun areContentsTheSame(oldItem: PopularKeywordData, newItem: PopularKeywordData): Boolean {
-            return oldItem == newItem
-        }
-    }
-}
-
-class SearchFeedListAdapter(
-    private val onClick: (SearchFeedData) -> Unit
-) : ListAdapter<SearchFeedData, SearchFeedListAdapter.SearchFeedViewHolder>(SearchFeedDiffUtil) {
-
-    class SearchFeedViewHolder(
-        private val binding: ItemPopularKeywordBinding,
-        private val onClick: (SearchFeedData) -> Unit
-    ) : RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(popularKeywordData: SearchFeedData) {
-            binding.root.setOnClickListener {
-                onClick(popularKeywordData)
-            }
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SearchFeedViewHolder {
-        val binding =
-            ItemPopularKeywordBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return SearchFeedViewHolder(binding, onClick)
-    }
-
-    override fun onBindViewHolder(holder: SearchFeedViewHolder, position: Int) {
-        val popularKeyword = getItem(position)
-        holder.bind(popularKeyword)
-    }
-
-    companion object SearchFeedDiffUtil : DiffUtil.ItemCallback<SearchFeedData>() {
-        override fun areItemsTheSame(oldItem: SearchFeedData, newItem: SearchFeedData): Boolean {
-            return oldItem.id == newItem.id
-        }
-
-        override fun areContentsTheSame(oldItem: SearchFeedData, newItem: SearchFeedData): Boolean {
-            return oldItem == newItem
-        }
-    }
-}
