+++ app/src/main/res/layout/item_gallery_album.xml
@@ -16,24 +16,24 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/title"
+        android:id="@+id/tv_folder_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="10dp"
         android:textColor="@color/white"
-        app:layout_constraintBottom_toTopOf="@id/subtitle"
+        app:layout_constraintBottom_toTopOf="@id/tv_folder_file_count"
         app:layout_constraintStart_toEndOf="@id/iv_album_image"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_chainStyle="packed" />
 
     <TextView
-        android:id="@+id/subtitle"
+        android:id="@+id/tv_folder_file_count"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="10dp"
         android:textColor="@color/blue_gray_4"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toEndOf="@id/iv_album_image"
-        app:layout_constraintTop_toBottomOf="@id/title" />
+        app:layout_constraintTop_toBottomOf="@id/tv_folder_name" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
