+++ app/src/main/res/layout/item_popular_keyword.xml
@@ -16,8 +16,19 @@
         android:padding="10dp">
 
         <TextView
+            android:id="@+id/tv_keyword_rank"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            tools:text="1"
+            android:textSize="14sp"
+            android:textColor="@color/white"
+            android:textStyle="bold" />
+
+        <TextView
+            android:id="@+id/tv_keyword"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="20dp"
             tools:text="비 오는 날"
             android:textSize="14sp"
             android:textColor="@color/white" />
