+++ app/src/main/res/layout/item_recent_keyword.xml
@@ -16,6 +16,7 @@
         android:padding="10dp">
 
         <TextView
+            android:id="@+id/tv_keyword"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             tools:text="물놀이"
